ai: Try to make Gemini CLI read both AGENTS.md and GEMINI.md... but it doesn't work.
According to https://github.com/google-gemini/gemini-cli/blob/main/docs/cli/gemini-md.md#customize-the-context-file-name,
this should work, but somehow it doesn't. I'll still commit this, in case it starts working in the future;
and in the meantime, just work-around it with e.g. https://github.com/enola-dev/enola/pull/1880 in projects.
(Doing it like that makes it work for other users as well - so that's probably better anyways.)